scheduler: initial commit
The scheduler is a singleton that manages running tasks, which may be marked
as either active or passive.  While there are active tasks, the scheduler is
active.  When all active and passive tasks are complete, the scheduler is
done.

When a task is added to the scheduler, the scheduler creates a context in
which to execute the task.  The context contains an initial ucontext and a
large amount of mmapped space for the stack.  The scheduler trampoline also
protects the rest of the process from C++ exceptions thrown in that context.

When a task is first scheduled, the context will be dequeued and the call
operator will be invoked.  If it schedules, a new, temporary ucontext is
created, usually on the stack, and that is scheduled.

When a task's call operator returns, the scheduler will reclaim the context
and switch to the next queued ucontext.  Tasks can also be owned by the
scheduler, in which case the scheduler deletes the task object.

The ucontext link pointer is used to maintain the intrusive_queue of running
ucontexts but can not be used to automatically switch ucontexts since that
apparently only works when makecontext is called.