reactor: initial commit
The reactor is a singleton that manages triggers, which wait_for activity on
a file_descriptor.  Under the hood, epoll is used.  A trigger owns the
file_descriptor on which it waits for activity and the FD can not be shared
between triggers; instead, the FD must be duped.

A trigger is armed by calling wait_for_* and then deferring the returned
ucontext from the scheduler.  Armed triggers can be disarmed by invoking
wait_for_nothing.  The returned ucontext should be referred to the
scheduler.

When the reactor is polled, it will attempt to drain as many events as
possible.  If wait is invoked instead, the reactor will block until at least
one event is ready and then poll.  The reactor then fires all armed triggers
for the drained events.  Note that an event might still be returned for a
trigger that has been disarmed.  These events are safely ignored.

When a trigger fires, the ucontext it owns is added to a context::queue; the
events in this queue should be referred to the scheduler soon.