signals: add suppressor
A signal_suppressor temporarily absorbs signals without disrupting any
pending signals at the time it is activated.  The intention is to prevent
SIGPIPE from causing the program to exit during write + friends; however, it
is general enough to work for any mergeable signals.

Since this project is intended to be used as a framework, it is not ideal to
just ignore SIGPIPE globally.  Instead, that signal is blocked temporarily.

If the signal was already pending, it should still be pending after the
suppressor is destroyed.  If it was not already pending, it should be
consumed.

The overall framework is still not signal safe at this point, since
switching ucontexts modifies the signal mask.  Therefore, signals should
only be blocked or unblocked outside of tasks.