concepts: killable + killer
A killable class responds to the kill method, after which it is dead.
Subclasses implement a die method, which will only be invoked the first time
an object is killed.

A killer targets killables, which it can kill in FILO order.  The targeted
object's lifetime should exceed that of the killer.