rfc862: add echo client and server
rfc862c is a client that connects to tcp:localhost:7, piping stdin to the
socket and piping the socket to stdout.

rfc862d launches an acceptor on tcp:7.  For each connection, it writes
everything it reads back into the socket.  The daemon handles SIGINT by
killing the acceptor but it will continue to run until all clients shutdown.

As a future improvement, it would be interesting to use splice instead of
reading to and writing from a buffer in the process.