core: initial commit
The core is a singleton that manages the main loop.  It wraps the scheduler
and the reactor in an event loop and keeps statistics.  Once tasks are
added, owned, or eyed by the core, react should be invoked.  Neither the
scheduler nor the reactor should be accessed directly.

All passive tasks being eyed are targeted by a killer.  When there are no
more active tasks, the killer kills all passive tasks.  The loop will
continue to run until no tasks remain.

The core abstracts certain operations as static methods.

 * schedule will create a new ucontext and place it onto a deferral
   queue, which will be transferred back to the scheduler after the next
   reactor poll.

 * read & write restart on interrupts, cause a trigger to defer to the
   reactor instead of blocking, and take a minimum number of bytes to
   process before returning.

 * shutdown is able to flush, deferring until the socket reports as not
   connected. Since only one side of the socket might be shutdown, we
   wait_for_error to avoid false-positives in the reactor.