task: connector
A simple_connector is a simple_task that connects to a socket_address.  It
has two callbacks, on_error and on_connect.

The connector<R,W> template class kills itself on error and creates new R
and W on connect, passing them FDs to own.  The new R and W are owned by
core.  The connector<T> template class is a specialization that only creates
a single new T.