task: acceptor
A simple_acceptor is a simple_task that binds to a socket_address and
listens.  It has two callbacks, on_error and on_accept.

The acceptor<T> template class kills itself on error and creates a new T on
accept, passing it an FD to own.  The new T is owned by the core.