task: signaler
A signaler is a simple_task that takes a signal_set and processes signals in
that set via signalfd.  It has one callback, on_signal, which takes a
signalfd_siginfo.

The signals in the signal_set passed to signaler should be blocked or else
they will still be handled normally.  Since core is not signal safe yet,
signals should not be blocked from running tasks.