intrusive: assert + set NULL link
It was lazy not to assert that T::*link is NULL when adding an object to an
intrusive structure and lazy not to set T::*link to NULL when removing an
object from an intrusive structure.

Originally, it seemed like any ucontext would transfer to context::uc_link
on termination even if uc_link was modified after makecontext.  It turns out
that this is not the case, so the scheduler must setcontext at the end of
the trampoline anyway and uc_link is meaningless except to the code that
reclaims context objects.

intrusive_stack just came along for intrusive_queue's misguided ride.