tree: replace predicate with key, add find
Originally, it seemed like a good idea to use a method pointer as an
ordering predicate.  This makes find awkward to implement and potentially
inconsistent with said predicate.

This change replaces that silliness with a member pointer and implements the
find operation in terms of it.  As a bonus, test_intrusive now includes an
invocation of find.