test: weaken assert for order + tree
While the assert is actually true, it is a bit too strong.  In general, with
repeated values tree::min might not be the same object that order::remove
returned, even though the values will always be the same.

In this case, it is true because both tree and order always err the same way
when they find an object of equal value.  However, that implementation
detail is subject to change.