rfc862c: add name resolution
Instead of just connecting to localhost, rfc862c now takes a list of hosts
to try on the command line.  It uses getaddrinfo to do the resolution, which
is not ideal in general but works well enough here.  On a failed connect, it
moves on to the next address and then the next host in the list.

After any successful connect, the process acts as an echo client (as before)
and then exits with EXIT_SUCCESS.  If there is no successful connection, it
exits with EXIT_FAILURE.