new: semaphore
This non-task class uses eventfd to act as a semaphore.  By default, it
contains one token, but can be given additional tokens on construction.  The
intended use is protecting large writes to shared resources, which may wait.

When invoking wait or signal, the semaphore attempts a read or a write; if
that fails, it then dups the trigger and waits.  This is due to the way
reactor uses epoll; only one task at a time may wait on a trigger, although
this is not properly reflected in the asserts in that class.

Because all waiting contexts are registered with the reactor, on a signal
they will all wake up but only one will actually be able to run.  It would
be nice if semaphore instead kept a queue of pending contexts waiting on the
semaphore and registered them on signal.

Additionally, it would be nice if semaphore implemented an RAII helper that
waited on construction and signaled on destruction.

These are both future improvements.