semaphore: add wait queue
This commit adds a wait queue to semaphore::wait, preventing unnecessary
duplication of eventfd FDs.  If the existing trigger is unarmed, then arm it
as usual.  If it is already armed, then enqueue a new context and defer.

When exiting semaphore::wait, if the trigger is armed then the triggered
context will bump the context::queue; otherwise, dequeue once and refer to
the scheduler.  This context will either read, re-arm, or defer again.  For
any signal, we will spuriously wake at most one extra waiting context.

In semaphore::signal, we can only get into trouble if the semaphore is too
large in the first place, so go ahead and dup the FD there.  For a binary
semaphore, this will never happen.