task: timer
A timer is a simple_task that handles timerfds with a fixed interval. It has
one callback, on_tick, which is invoked every time the timer goes off.  If
multiple timer events happen between reads, multiple ticks may occur due to
a single read.

For variable timers or other delays, more appropriate infrastructure is
required, likely in the reactor itself.