library: use lace + lite
The utility code and intrusive templates aren't specific to liboco and the
intention to move them elsewhere has existed for quite a while.  This change
does that and now requires libace and libite to be installed.  Additionally,
namespace and file name changes demanded light editing.

Now that liboco is using libite, it will be much easier to take advantage of
other intrusive containers in the library.  A likely first project is the
addition of core::sleep via a heap in the reactor.