scheduler: s/MINSIGSTKSZ/SIGSTKSZ/ for trampoline stack
This should not have been MINSIGSTKSZ, since that's a minimum signal stack
size --- plus this code doesn't have anything to do with stacks anyway.
Since SIGSTKSZ is considered a "reasonable" stack size for handling, it
probably makes sense to use it here instead.

The scheduler should not actually be handling signals inside of the ucontext
trampoline anyway.  In general, important signals will be masked and handled
by a signalfd.  A future change will set an alternate signal stack inside of
the invocation of core::react.