core: (ab)use signal stack cookie to implement current
This change is the first step in making core not a singleton.  On a given
thread, only one core can react at a time since all of the management data
structures are lockless.  Therefore, at the beginning of react, set cookie
to the current core and use sigaltstack ::current to retrieve the cookie in
core::current.

This method must not be invoked outside of react, i.e. outside a currently
active core.

Future changes will make core own a queue, reactor, & scheduler rather than
using singleton instances of those classes.