trigger: lazy add to reactor
In anticipation of non-singleton cores owning reactors directly, lazy add to
the reactor.  In the future, triggers will need to assert that the cached
reactor is the same as the current reactor and either assert that or change
to a different reactor in response on the assumption that the old core &
reactor are dead.  Given object lifetimes, this might not be so trivial
unless we can guarantee that the lifetime of all reactors is longer than all
triggers.

Since reactor is still a singleton, we need to grab an instance in the
trigger constructor in order to ensure that the reactor is constructed
before and destructed after all the triggers that are attached to it.
Otherwise, the reactor might not be alone during destruction and will
assert.