reactor: remove total triggers counter
In the mysterious future, where reactor is not a singleton and therefore its
lifetime is uncertain with respect to its registered tiggers, it would be
nice if trigger did not have to remove itself from the reactor.  With epoll,
this is automatic on file descriptor close.  However, it means that the
reactor might die with a non-zero trigger count and that lingering triggers
might try to access the reactor after destruction.  Therefore, remove
trigger count and do not assert that the reactor is alone.

This is not ideal, but the alternative is to have triggers hang around in a
list in the reactor and unregister them if they are still alive during
reactor destruction.  This might be necessary if tasks are able to change
cores, which implies that triggers could change reactors.  For now, just let
the reactor implicitly remove trigger file descriptors and when reactor is
not a singleton assert that the current singleton matches the registered
reactor.

The important assertion is that reactor has no armed triggers, since it
means they will never fire and will therefore hang.

It would also be nice if reactor handled all the details for epoll.